+++ Session-09/CalculatorLib/Subtraction.cs
@@ -6,8 +6,8 @@ using System.Threading.Tasks;
 
 namespace CalculatorLib {
     public class Subtraction {
-        public decimal Do(decimal? x, decimal? y) {
-            decimal ret = 0;
+        public double Do(double? x, double? y) {
+            double ret = 0;
             if (x != null && y != null) {
                 ret = x.Value - y.Value;
             }
